function opinter

+++ operators.c
@@ -1,12 +1,12 @@
 #include "operators.h"
 
-int add(int op1, int op2) {
+double add(int op1, int op2) {
 	return op1+op2;
 }
-int minus(int op1, int op2) {
+double minus(int op1, int op2) {
 	return op1-op2;
 }
-int mul(int op1, int op2) {
+double mul(int op1, int op2) {
 	return op1*op2;
 }
